factor `ElementTypeHint` into its own file
